mpv-Subtitle-Definition/requirements.txt
openai
nltk
python-dotenv